+++ internal/_internals.ts
@@ -1,0 +1,4 @@
+import { diff } from "./diff.ts";
+import { diffstr } from "./diff_str.ts";
+
+export const _internals = { diff, diffstr };
