+++ internal/mod.ts
@@ -7,5 +7,8 @@
  *
  * @module
  */
-export * from "./format.ts";
+export * from "./_internals.ts";
+export * from "./build_message.ts";
 export * from "./diff.ts";
+export * from "./diff_str.ts";
+export * from "./format.ts";
