+++ internal/types.ts
@@ -1,0 +1,7 @@
+export type DiffType = "removed" | "common" | "added";
+
+export interface DiffResult<T> {
+  type: DiffType;
+  value: T;
+  details?: Array<DiffResult<T>>;
+}
